+++ Action.ts
@@ -10,18 +10,13 @@ import { Room, Client } from "colyseus";
 //     }
 // }
 
-// Allows objects other than the Room to create callback functions that utilize Room actions
-
-export type CallbackFunction = (...args: any[]) => void;
-export const emptyCallback: CallbackFunction = function() {};
-
-export type ActionFunction = (client: Client, params: any, Room: Room) => CallbackFunction[];
+export type ActionFunction = (client: Client, params: any, Room: Room) => void;
 
 // Not required, just handy
 
 export enum actions {
     setPlayerName = "setPlayerName",
     updateSelectedRole = "updateSelectedRole",
-    startGame = "startGame"
+    startGame = "startGame",
 }
 export type Action = keyof typeof actions;
